reorganize crate modules
move to newer style (removing mod.rs files)
update usings to properly prefix modules for imported types and methods
move code around and rename several types for more clarity in the code

this work is not finished but this gets us much closer to where we want to be
