+++ src/test/template.rs
@@ -1,15 +1,14 @@
-use crate::test::definition::{Modifier, ResponseExtraction, VariableTypes};
-use crate::test::file::{
-    TestFile, UnvalidatedCleanup, UnvalidatedCompareRequest, UnvalidatedRequest,
-    UnvalidatedRequestResponse, UnvalidatedResponse, UnvalidatedStage, UnvalidatedVariable,
-};
-use crate::test::http::{HttpHeader, HttpParameter, HttpVerb};
+use crate::test;
+use crate::test::definition;
+use crate::test::file;
+use crate::test::http::{Header, Parameter, Verb};
+use crate::test::variable;
 use std::cell::Cell;
 use std::error::Error;
 use uuid::Uuid;
 
-pub fn template() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
-    Ok(TestFile {
+pub fn template() -> Result<test::File, Box<dyn Error + Send + Sync>> {
+    Ok(test::File {
         filename: "".to_string(),
         name: Some("".to_string()),
         id: Some(Uuid::new_v4().to_string()),
@@ -27,8 +26,8 @@ pub fn template() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
     })
 }
 
-pub fn template_staged() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
-    Ok(TestFile {
+pub fn template_staged() -> Result<test::File, Box<dyn Error + Send + Sync>> {
+    Ok(test::File {
         filename: "".to_string(),
         name: Some("".to_string()),
         id: Some(Uuid::new_v4().to_string()),
@@ -46,8 +45,8 @@ pub fn template_staged() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
     })
 }
 
-pub fn template_full() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
-    Ok(TestFile {
+pub fn template_full() -> Result<test::File, Box<dyn Error + Send + Sync>> {
+    Ok(test::File {
         filename: "".to_string(),
         name: Some("".to_string()),
         id: Some(Uuid::new_v4().to_string()),
@@ -65,23 +64,24 @@ pub fn template_full() -> Result<TestFile, Box<dyn Error + Send + Sync>> {
     })
 }
 
-fn new_full_cleanup() -> Result<UnvalidatedCleanup, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedCleanup {
+fn new_full_cleanup() -> Result<file::UnvalidatedCleanup, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedCleanup {
         request: Some(new_full_request()?),
         onsuccess: Some(new_full_request()?),
         onfailure: Some(new_full_request()?),
     })
 }
 
-fn new_full_request_response() -> Result<UnvalidatedRequestResponse, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedRequestResponse {
+fn new_full_request_response(
+) -> Result<file::UnvalidatedRequestResponse, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedRequestResponse {
         request: new_full_request()?,
         response: Some(new_full_response()?),
     })
 }
 
-fn new_stage() -> UnvalidatedStage {
-    UnvalidatedStage {
+fn new_stage() -> file::UnvalidatedStage {
+    file::UnvalidatedStage {
         request: new_request(),
         compare: None,
         response: Some(new_response()),
@@ -89,8 +89,8 @@ fn new_stage() -> UnvalidatedStage {
     }
 }
 
-fn new_full_stage() -> Result<UnvalidatedStage, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedStage {
+fn new_full_stage() -> Result<file::UnvalidatedStage, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedStage {
         request: new_full_request()?,
         compare: Some(new_full_compare()?),
         response: Some(new_full_response()?),
@@ -98,18 +98,18 @@ fn new_full_stage() -> Result<UnvalidatedStage, Box<dyn Error + Send + Sync>> {
     })
 }
 
-fn new_full_variable() -> Result<UnvalidatedVariable, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedVariable {
+fn new_full_variable() -> Result<file::UnvalidatedVariable, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedVariable {
         name: "".to_string(),
-        data_type: VariableTypes::String,
+        data_type: variable::Type::String,
         value: serde_json::from_str("{}")?,
-        modifier: Some(Modifier::new()),
+        modifier: Some(variable::Modifier::new()),
         format: Some("".to_string()),
     })
 }
 
-fn new_response() -> UnvalidatedResponse {
-    UnvalidatedResponse {
+fn new_response() -> file::UnvalidatedResponse {
+    file::UnvalidatedResponse {
         status: Some(200),
         headers: None,
         body: None,
@@ -118,19 +118,19 @@ fn new_response() -> UnvalidatedResponse {
     }
 }
 
-fn new_full_response() -> Result<UnvalidatedResponse, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedResponse {
+fn new_full_response() -> Result<file::UnvalidatedResponse, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedResponse {
         status: Some(200),
         headers: Some(vec![new_header()]),
         body: Some(serde_json::from_str("{}")?),
         ignore: Some(vec!["".to_string()]),
-        extract: Some(vec![ResponseExtraction::new()]),
+        extract: Some(vec![definition::ResponseExtraction::new()]),
     })
 }
 
-fn new_full_compare() -> Result<UnvalidatedCompareRequest, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedCompareRequest {
-        method: Some(HttpVerb::Get),
+fn new_full_compare() -> Result<file::UnvalidatedCompareRequest, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedCompareRequest {
+        method: Some(Verb::Get),
         url: "".to_string(),
         params: Some(vec![new_parameter()]),
         add_params: Some(vec![new_parameter()]),
@@ -142,8 +142,8 @@ fn new_full_compare() -> Result<UnvalidatedCompareRequest, Box<dyn Error + Send 
     })
 }
 
-fn new_request() -> UnvalidatedRequest {
-    UnvalidatedRequest {
+fn new_request() -> file::UnvalidatedRequest {
+    file::UnvalidatedRequest {
         method: None,
         url: "".to_string(),
         params: None,
@@ -152,9 +152,9 @@ fn new_request() -> UnvalidatedRequest {
     }
 }
 
-fn new_full_request() -> Result<UnvalidatedRequest, Box<dyn Error + Send + Sync>> {
-    Ok(UnvalidatedRequest {
-        method: Some(HttpVerb::Get),
+fn new_full_request() -> Result<file::UnvalidatedRequest, Box<dyn Error + Send + Sync>> {
+    Ok(file::UnvalidatedRequest {
+        method: Some(Verb::Get),
         url: "".to_string(),
         params: Some(vec![new_parameter()]),
         headers: Some(vec![new_header()]),
@@ -162,8 +162,8 @@ fn new_full_request() -> Result<UnvalidatedRequest, Box<dyn Error + Send + Sync>
     })
 }
 
-fn new_header() -> HttpHeader {
-    HttpHeader {
+fn new_header() -> Header {
+    Header {
         header: "".to_string(),
         value: "".to_string(),
 
@@ -171,8 +171,8 @@ fn new_header() -> HttpHeader {
     }
 }
 
-fn new_parameter() -> HttpParameter {
-    HttpParameter {
+fn new_parameter() -> Parameter {
+    Parameter {
         param: "".to_string(),
         value: "".to_string(),
