--- src/test/validator.rs
@@ -1,1 +1,0 @@
-
